Create new interface to GA toolbox called mpga.m
This commit represents multiple years of modifications the GA toolbox to improve it's interface. Primarily it adds mpga.m, an interface function to the toolbox which, among other things facilitates passing additional arguments to the objective function and other functions. It also improves reporting of the results and allows you to stop the GA, save progress and restart the GA from where you left off. It also provides an easy way to regenerate an individual from a chromosome, to get the real results of an evaluation rather than just the score.